# src/docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "TV Library Backend",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation for TV Library Server based on the OpenAPI 3.0 specification. \nSome useful links:\n- [TV Library Backend repository](https://github.com/DmytriiTsybuliak/movies-library-backend)\n- [The source API definition for the TV Library Backend](https://github.com/DmytriiTsybuliak/movies-library-backend/src/docs/openapi.yaml)"
  },
  "servers": [
    {
      "url": "https://movies-library-backend-s1fd.onrender.com"
    }
  ],
...
  "paths": {
...
    "/auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset password",
        "description": "",
        "operationId": "resetPassword",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "jwt token from the link of email"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "password"
                ],
                "properties": {
                  "password": {
                    "type": "string",
                    "example": "12345678",
                    "description": "new password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Password was successfully reset!"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User with this email not found / New password must not match the old one",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Something went wrong"
                    },
                    "error": {
                      "type": "string",
                      "example": "User with this email not found"
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionId"
      }
    },
...
  }
}